refactor(Client): Separate api and client modules (#2199)
# Description

- The `client` module contains the new `Argilla` class (old `Api`) 
- The `api` module set up the main argilla functions like `load` and
`log`.

Even if both modules shares docstring and method signature, I prefer to
handle them separately. This will simplify when argilla client instances
are exposed for better integrations.

Closes #2194 

**Type of change**

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I added comments to my code
- [x] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
