+++ src/argilla/client/sdk/commons/errors.py
@@ -31,6 +31,10 @@ class WrongResponseError(BaseClientError):
         )
 
 
+class InputValueError(BaseClientError):
+    pass
+
+
 class ApiCompatibilityError(BaseClientError):
     def __init__(self, min_version: str):
         self.min_version = min_version
