+++ tests/client/test_init.py
@@ -12,6 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
+from argilla import TextClassificationRecord
 from argilla.client import api
 
 
@@ -26,7 +27,9 @@ def test_resource_leaking_with_several_init(mocked_client):
     for i in range(0, 10):
         api.init()
         api.log(
-            api.TextClassificationRecord(text="The text"), name=dataset, verbose=False
+            TextClassificationRecord(text="The text"),
+            name=dataset,
+            verbose=False,
         )
 
     assert len(api.load(dataset)) == 10
