+++ tests/functional_tests/test_log_for_token_classification.py
@@ -59,7 +59,7 @@ def test_call_metrics_with_no_api_client_initialized(mocked_client):
         if metric == entity_consistency:
             continue
 
-        api.__ACTIVE_API__ = None
+        api.ArgillaSingleton.clear()
         with pytest.raises(NotFoundApiError):
             metric("not_found")
